+++ src/traits/option_ext.rs
@@ -1,5 +1,7 @@
+
 use crate::traits::string_ext::StrLowerExt;
 
+#[allow(dead_code)]
 pub trait OptionExt {
     type Inner;