+++ src/traits/vec_ext.rs
@@ -2,6 +2,7 @@
 
 /// This trait to add helper functions to Vec<String> for ease of use in searching ignoring case
 
+#[allow(unused_variables)]
 pub trait VecStringExt {
     fn contains_ignore_case(&self, query: &str) -> bool;
     fn contains_any(&self, queries: &[&str]) -> bool;
@@ -13,6 +14,7 @@ impl VecStringExt for Vec<String> {
         self.iter().any(|s| s.to_lowercase().contains(&query_lower))
     }
 
+    #[allow(unused_variables)]
     fn contains_any(&self, queries: &[&str]) -> bool {
         self.iter().any(|q| self.contains_ignore_case(q))
     }